+++ src/backend/joanie/tests/core/api/order/test_submit_for_signature.py
@@ -9,6 +9,7 @@ from django.urls import reverse
 from django.utils import timezone as django_timezone
 
 from joanie.core import enums, factories
+from joanie.payment.factories import InvoiceFactory
 from joanie.tests.base import BaseAPITestCase
 
 
@@ -149,12 +150,12 @@ class OrderSubmitForSignatureApiTest(BaseAPITestCase):
         user = factories.UserFactory(
             email="student_do@example.fr", first_name="John Doe", last_name=""
         )
-        factories.AddressFactory(owner=user)
         order = factories.OrderFactory(
             owner=user,
             state=enums.ORDER_STATE_VALIDATED,
             product=factories.ProductFactory(),
         )
+        InvoiceFactory(order=order)
         token = self.get_user_token(user.username)
         expected_substring_invite_url = (
             "https://dummysignaturebackend.fr/?requestToken="
@@ -195,12 +196,12 @@ class OrderSubmitForSignatureApiTest(BaseAPITestCase):
         user = factories.UserFactory(
             email="student_do@example.fr", first_name="John Doe", last_name=""
         )
-        factories.AddressFactory(owner=user)
         order = factories.OrderFactory(
             owner=user,
             state=enums.ORDER_STATE_VALIDATED,
             product=factories.ProductFactory(),
         )
+        InvoiceFactory(order=order)
         token = self.get_user_token(user.username)
         contract = factories.ContractFactory(
             order=order,
@@ -247,12 +248,12 @@ class OrderSubmitForSignatureApiTest(BaseAPITestCase):
         user = factories.UserFactory(
             email="student_do@example.fr", first_name="John Doe", last_name=""
         )
-        factories.AddressFactory(owner=user)
         order = factories.OrderFactory(
             owner=user,
             state=enums.ORDER_STATE_VALIDATED,
             product=factories.ProductFactory(),
         )
+        InvoiceFactory(order=order)
         token = self.get_user_token(user.username)
         contract = factories.ContractFactory(
             order=order,